fix: segmented download broken for files > 2GB due to Int overflow
SegmentCalculator.calculateSegments() used totalBytes.toInt() which
overflows for files exceeding Int.MAX_VALUE (2,147,483,647 bytes).
This caused the effective connection count to collapse to 1, resulting
in a single segment covering the entire file.

Fix: compare using Long arithmetic before converting to Int.

+++ library/core/src/commonMain/kotlin/com/linroid/kdown/core/segment/SegmentCalculator.kt
@@ -8,7 +8,8 @@ internal object SegmentCalculator {
     require(totalBytes > 0) { "totalBytes must be positive" }
     require(connections > 0) { "connections must be positive" }
 
-    val effectiveConnections = minOf(connections, totalBytes.toInt().coerceAtLeast(1))
+    val effectiveConnections =
+      minOf(connections.toLong(), totalBytes.coerceAtLeast(1)).toInt()
     val segmentSize = totalBytes / effectiveConnections
     val remainder = totalBytes % effectiveConnections
